Feat: 将 User 分为四类，使用 4 个 Moonlark User 子类实现
