+++ src/plugins/nonebot_plugin_larkuid/web/user.py
@@ -6,8 +6,7 @@ from nonebot import get_app
 from ...nonebot_plugin_larkuser.utils.user import get_user
 from ..types import BasicUserResponse, DetailedUserResponse
 from ..session import get_user_data
-from ...nonebot_plugin_larkuser import MoonlarkUser
-
+from ...nonebot_plugin_larkuser.user.base import MoonlarkUser
 
 app = cast(FastAPI, get_app())
