+++ src/plugins/nonebot_plugin_larkuser/__init__.py
@@ -19,4 +19,5 @@ require("nonebot_plugin_render")
 from .matchers import recorder, panel, whoami
 from .utils.matcher import patch_matcher
 from .utils.level import get_level_by_experience
-from .utils.user import get_user, MoonlarkUser
+from .utils.user import get_user
+from .user.base import MoonlarkUser
