+++ src/plugins/nonebot_plugin_sign/__main__.py
@@ -17,7 +17,8 @@ from ..nonebot_plugin_render.render import render_template
 from ..nonebot_plugin_email.utils.unread import get_unread_email_count
 from ..nonebot_plugin_jrrp.jrrp import get_luck_value
 from src.plugins.nonebot_plugin_larkuser.utils.matcher import patch_matcher
-from ..nonebot_plugin_larkuser import MoonlarkUser, get_user
+from ..nonebot_plugin_larkuser import get_user
+from ..nonebot_plugin_larkuser.user.base import MoonlarkUser
 from ..nonebot_plugin_larkutils import get_user_id
 from .config import config
 from .lang import lang
